% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aFancyFunction.R
\name{aFancyFunction}
\alias{aFancyFunction}
\title{A Fancy Function}
\usage{
aFancyFunction()
}
\value{
No return value.
}
\description{
This is my first function and it prints "This is fancy!"
}
\examples{
aFancyFunction()
}
