% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFanciness.R
\name{checkFanciness}
\alias{checkFanciness}
\title{A Function that Checks Fanciness}
\usage{
checkFanciness(check)
}
\arguments{
\item{check}{The term to be checked for fanciness.}
}
\value{
No return value.
}
\description{
This is my second function and it prints the level of fanciness of a given term.
}
\examples{
checkFanciness(3)
}
