% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyEnough.R
\name{fancyEnough}
\alias{fancyEnough}
\title{A Function that Checks Whether Something is Fancy Enough}
\usage{
fancyEnough(check, threshold)
}
\arguments{
\item{check}{The term to be checked.}

\item{threshold}{The required fanciness threshold.}
}
\value{
`TRUE` if `check` is fancy enough and `FALSE`if `check` is not fancy enough.
}
\description{
This is my third function and it checks whether a given term is fancy enough.
}
\examples{
fancyEnough(3,1)
fancyEnough(3,4)
}
